requirements.txt
selenium
Appium-Python-Client
PyYAML
pytest
allure-pytest
flaky
pytest-sugar
jprops
allure-python-commons
uiautomator2